+++ lib/ember/appkit/rails/active_support.rb
@@ -1,0 +1,1 @@
+require 'ember/appkit/rails/active_support/dependencies'